+++ charts/observability-pipelines-worker/CHANGELOG.md
@@ -1,5 +1,13 @@
 # Changelog
 
+## 2.1.2
+
+* Official image `2.1.2`
+
+## 2.1.1
+
+* Official image `2.1.1`
+
 ## 2.1.0
 
 * Official image `2.1.0`