+++ charts/synthetics-private-location/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Datadog changelog
 
+## 0.17.1
+
+* Update private location image version to `1.50.0`.
+
 ## 0.17.0
 
 * Add `podDisruptionBudget` to allow creating and configuring PodDisruptionBudget for deployment.